CommandList key 제한 변경

+++ MannequinHunter/Source/MannequinHunter/Public/AI/AIPattern.h
@@ -51,7 +51,7 @@ struct MANNEQUINHUNTER_API FAIPatternTable : public FTableRowBase
 	GENERATED_BODY()
 
 	UPROPERTY(EditDefaultsOnly)
-	TMap<float, FAIPatternDataArray> patternData;
+	TMap<int32, FAIPatternDataArray> patternData;
 
 };
